Initial Commit
Created the chessboard, mapped the pieces to 2d array and rendered them to the screen. Implemented anarchy like movement where pieces can move to any square. Model View Controller is the framework used to handle the project. The logic of the pieces moving is handled within the array and the a simple render function deals with the view side of things.